feat: 뷰포트는 픽셀 단위로 계산, 기즈모 스케일 조정

+++ EngineSIU/EngineSIU/Engine/Source/Runtime/Slate/Widgets/Layout/SSplitter.h
@@ -14,7 +14,7 @@ public:
 
     uint32 SplitterLimitLT = 100; // Pixel Value
 
-    float SplitterHalfThickness = 5.f; // Pixel Value
+    uint32 SplitterHalfThickness = 5; // Pixel Value
 
     bool bIsSplitterPressed = false;
